Create timer_lcd.c
Program that displays time on the first row of LCD display in 1 second increments (00 to 99). On the second line of the LCD display it must display two quantities. The first one is the time in increment of 7 seconds where in every 7 seconds the LED’s connected to PORTC will turn on for 5 seconds and then turn off. The second quantity is the time in increment of 4 seconds where in every 4 seconds the LED’s connected to PORTB turn on for 2 seconds and LED’s connected to PORTD turn on for 3 seconds.
The LCD should look something like the following:
TIME IN SEC: 45
T7: 14   T4: 16                 

All times must rollover when they reach 99.